backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-multipart>=0.0.12
httpx>=0.27.0
numpy>=2.1.0
scikit-learn>=1.5.0
python-dotenv>=1.0.0
email-validator>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
networkx>=3.3
